Reduce over linking of terms in README
Wikipedia's linking guidelines[1] say, "Generally, a link should appear
only once in an article…" It is useful for terms to be linked again
in info boxes and other sections, but the README currently links terms
multiple times in normal phrasing.

Removing some of the links ensures only new ideas and terms are
highlighted in blue.

[1] https://en.wikipedia.org/wiki/Wikipedia:Manual_of_Style/Linking#What_generally_should_not_be_linked