Send only editable attributes when updating an App
Backbone sends all model attributes by default when a model is saved. In
an App's case, that includes immutable values like "tasksRunning" and
"tasksStaged" and the "version" value, which prevents any change from
happening.

Add a "editable attributes" list that is a whitelist of attributes that
should be sent to the server on save.

Fixes #198.