Allow leading slash in app names in UI
If an app name does not start with a slash when posted to the API, the
API prepends a slash and returns that name. The UI expected app names to
start only with alphanumeric characters and so raised a validation error
when the API responded with the new name.

Let the app name regex start with an optional slash so the UI doesn’t
require it during creation but will allow it when the name is returned
from the API.

Fixes #445.