Remove incorrect `cmd` validation from frontend
The dependency of `cmd` and `container` is not defined in the backend is
now incorrect in the frontend. Rather than re-implement validation here,
defer to the backend for complex inter-field dependencies.

The `container` format changed, will also prevented scaling and
suspending from the UI. This fixes scaling/suspending in the UI for apps
using the new container format.

Fixes #496.