Reconcile health checks after a task is launched
During scale up new tasks are started with a new appVersion. Because each health
check only checks the tasks of exactly one appVersion, the health checks before
the scaling ignore the new tasks.

New health checks are started by the HealthCheckManager on "reconcileWith" which
before this patch was called only after app creation and on leader change. This
patch adds another call after a task is launched and after scale down.

Note, that it would not be enough to call reconcileWith after
MarathonSchedulerActor.scale on scale up because new tasks are launched
asynchronously and reconcileWith would not see new tasks yet.

Fixes #901.