Don't create unknown tasks with incomplete data in TaskTracker.running
When a task is not known by the TaskTracker when TaskTracker.running is
called by the scheduler, before this patch an incomplete task was stored
in the TaskTracker, recognizable from the 1970 beginning of epoch date.

This somehow happens sometimes after reconciation, wenn Mesos sends
status updates of task not in the task store after some leader change.

The actual bug is that those tasks are missing in the task store. An
unproven theory is that health checks run by non-leaders are remove
tasks from the store although after leader defeat all health checks must
be stopped. But this is not true right now, compare #356.

Workaround for #896.