Update app version of old tasks after scaling
When an app is scaled with existing tasks, the version of the new and of the old
tasks was different. This looked confusing in the UI for the user.

This patch updates all the old tasks at the end of the scaling deployment step.

Backend part of the fix for #1025