Catch exceptions during election correctly and offer leadership again
The previous patch did not fix the problem of a
java.util.concurrent.TimeoutException exception thrown by the Mesos libs during
migration because this exception is not of type mesosphere.marathon.Exception.

When changing the type in the catch clause to scala.Exception the catch works,
but because the driver is not running yet, no offerLeadership is called.

This patch calls the abdication command and offerLeadership explicitly if the
driver is not yet running.

Moreover, a offerLeadership backoff mechanism is implemented because otherwise
very fast repeated election can happen when a host is bad.

Finally, a test case is added for the java.util.concurrent.TimeoutException
case in Migration.migrate.

Fixes #1043