Turn Timestamp into a normal case class and inherit equal and hashCode
Timestamp was a case class with overridden `equal` and `hashCode` methods which
used the UTC normalized dateTime instead of the parameter dataTime.

This patch removes the `time` value inside the case class. It enforces UTC
in the parameter `dateTime` by hiding the constructor and adding timezone
conversion to the companion object's `apply(dateTime: DateTime)`.

The Timestamp case class is abstract (following https://issues.scala-lang.org/browse/SI-844)
such that the definition of `apply` is omitted by the compiler. Hence, we can
define our own `apply` in the companion option without conflict.