Log "No matching offer ..." at level WARN
Logging at INFO in a production environment can be rather verbose, but it is important to capture "No matching offer ..." messages, since it can indicate issues with an application configuration, or other issue with the cluster. This change moves the "No matching offer ..." log message from the INFO level to the WARN level to make it more visible.