Fixes #1573 - Report number of apps/groups in metrics
Added the number of add/groups to the metrics exposed by Marathon.

Note: With this change, Marathon will load all the groups from ZK each
time the /metrics endpoint is hit. If this becomes a performance
issue, we can cache the metrics and move the refreshing logic to an
actor. I decided to try to avoid premature optimizations.