Improves #1563 - Move help resources into doc directory
to be able to include them in the rest-api.md.

Since we have to consolidate the information from rest-api.md and the
help pages (both potentially contain updated information not present in the other one),
I only added two examplary includes.

I needed to update Jekyll to be able to use `include_relative`. I also updated other
tools by executing `bundle update` since sorting out each dependency update that
is necessary for a newer Jekyll version would be more work.

I already prepared the markup for compatability by replacing the {% highlight %}
syntax with the one for markdown (```http ... ```)

The rest of the work will be performed in separate PRs.