Update to visualiser v 1.5.1
This integrates parts of Matt's feedback from previous release.
The biggest change is the replacement of the staged containers counter
with a stopwatch and a peak containers per second indicator. This
measures the highest number of running containers per second as
received by the API.

UPDATE: the stopwatch works againsta "target" value which is determined
by the sum of the instances number for each app definition.

In this release a capture.sh script has been added, together with
instructions in the README. This is useful for testing.

- Replaced the total staged containers indicator with a timer
- Verizon logo added
- Increased timeout value for ajax requests. This is a possible cause
for what Craig reported as "lost" increments in the running indicator.
You can track this down in the Slack verizon-mesosphere channel.