Towards #3041 - Introduce a non-protocol-buffer Task class
* extend MarathonTask with volume reservation marker and launch counter
* create Task class and serialization from/to MarathonTask protocol
  buffer objects

A lot of tests broke because they used unrealistic task data which is not
representable by the new Task anymore.

This is the first step of a gradual replacement of MarathonTask with TaskState
everywhere except inside of the task.tracker package for serialization.

Also remove MarathonTestHelper trait to make incremental compiles much faster
and not confuse Intellij that much.