Fixes #3438 - Integration Tests
Prior integration tests depended on restart-on-failure logic.
Unfortunately, that masked a race-condition in Mesos. I tried to
unmask as much as possible without getting failing tests.

In addition, we know check for a clean slate in Mesos after/before
every test.

There is also one test that checks for the expected reservations
for a resident app.