Support for creating persistent volumes on path or mount disk resources
Tries and fulfill largest disk resource requests, first. This could help Marathon to do a more intelligent thing and not allocate
itself into a corner. Given the following two resources:

- path resource with 128m free
- path resource with 628m free

If we had a task attempting to allocate two pesistent disks, one for
100m, and the other for 600m, it would fail if the 100mb were subtracted
from the 628m resource, first (which is the default
precedence for path disk resources, to evenly distribute free space).