Refactored 'GroupManager' as an actor. (#4098)
* Refactored 'GroupManager' as an actor.

Following the pattern in the `mesosphere.marathon.core` namespace,
'GroupManager' is refactored as a module. The existing behavior hasn't been
changed and leave room for further refactorings. While 'GroupManager' is now an
actor, it doesn't persist any state yet. Further work could focus on that.

* Resolve a cyclic dependency issue of Guice.

A 'Provider' is used to break the cycle. When MarathonSchedulerService
instanciation is moved to CoreGuiceModule, this provider could be removed.

* Refactored actor message handling.

* Improved the GroupManager configuration and renamed the package.

* Documented response types.