Disable the Mesos containerizer integration test.
At the moment, the integration tests are run inside a Docker container.
This poses some challenges for the Mesos containerizer. To be able to
use it with Docker images, it needs read/write access to cgroups, which
is hard to do inside a container. For now the test has been disabled
until a solution is found.