Don't consider lost tasks when evaluating constraints.
This change should let Marathon start a new task to replace a lost one,
even if an app is pinned to a slave. This could potentially lead to more
than one task running on the same agent, violating a UNIQUE constraint.
Marathon will in that case try to remediate it by killing the new
running/staging task.