Me/cherry pick 0.15 lost handling to 1.3 (#4307)
Expunge lost tasks if the task update message indicates it won't come back (#4241)
Don't consider lost tasks when evaluating constraints. (#4242)

This change should let Marathon start a new task to replace a lost one,
even if an app is pinned to a slave. This could potentially lead to more
than one task running on the same agent, violating a UNIQUE constraint.
Marathon will in that case try to remediate it by killing a staging task, if all
tasks are running, it will kill the oldest one.