Fix broken HealthStatus reporting for pod instances.
The logic to infer instance health changes was broken in that it only reported instances as healthy if all tasks were reported healthy. This is wrong when a task has no configured healthCheck and will never have a health status.

This patch changes this to

* an instance is healthy if
  all tasks are running or finished AND
  no task is reported unhealthy AND
  at least one task is reported healthy

* an instance is unhealthy if
  at least one task is running and reported unhealthy (this ignores unhealthy tasks which are not running)

* an instance has no health status if
  not all tasks are running or finished OR
  no task has health information

This should only act as a band aid until we provide information about the actual configured health checks. (An instance should not be reported healthy if a task that has a healthCheck configured has no health status information yet.)