Fix task state transition for terminal tasks
Any processed TaskUpdateOperation for a terminal task will now lead to a Noop. This is to prevent bugs when Marathon tries to kill a task that has already terminated. It will likely receive a TASK_LOST as a result, but the task is actually not lost but still e.g. Finished.