Don't wait for kills to finish.
Summary:
This avoids time outs as reported in DCOS-10599. Kills can take a long
time to complete so we opt for a fire and forget. Users of the API can
poll to see if a task was killed.

This is not a fix or reversal for #4191.

Test Plan: unit tests

Reviewers: aquamatthias

Reviewed By: aquamatthias

Subscribers: marathon-team, jenkins

Differential Revision: https://phabricator.mesosphere.com/D72