Fix deployment step timeout issue that could cause a deployment failure
Summary:
A timeout in the TaskStartActor (or other deployment step actors) may cause
the whole deployment to fail because it's handling is not separated from
the actor stop (e.g. in case of a master abdication).
- A separate Shutdown message is sent to handle master abdication that will try
to fail the corresponding step promise
- TaskStartActor (and other deployment step actors) should be able now to simply
restart on exception (by default supervision strategy)

Related: DCOS-10200

Test Plan: manual test

Reviewers: unterstein, jdef

Reviewed By: unterstein, jdef

Subscribers: jdef, jasongilanfarr, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D88