fix instance zero-value generator for legacy store compat
Summary:
Legacy store codepath is very unhappy when loading existing instances/tasks because of a conflict with the zero-value generator and the regular expression we use to validate instance IDs.
It's impacting my migration testing, but I suspect would be an issue for anyone who tries to migrate to 1.4 using the legacy store with tasks in-flight.

Test Plan: sbt test

Reviewers: meichstedt, unterstein, jasongilanfarr

Reviewed By: jasongilanfarr

Subscribers: jenkins, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D109