Fix a questionable usage of iterator.filter.map
Summary:
DCOS-10951 has a massive stack trace that indicates we are
somehow calling filter.map on an iterator in a way that gives
a StackOverflow. This is the only call to ".filter" I could
find that was (a) in an Actor, (b) on an Iterator, and (c) is followed
by a call to ".map".

Test Plan: sbt test

Reviewers: meichstedt, jdef

Reviewed By: jdef

Subscribers: marathon-team

Differential Revision: https://phabricator.mesosphere.com/D110