Fixed a bug where a non-root `Group` can be passed to `DeploymentPlan` in dry-run mode.
Summary:
This was found during the `GroupTree` refactoring work.
`DeploymentPlan` expects root groups to be passed to it, since we invoke
`dependenGraph` on it, but in dry-run mode we were passing non-root groups
in cases where a valid non-root group was specified in the API.

Test Plan: set test + added a test which breaks without the changes made in this patch.

Reviewers: jasongilanfarr, aquamatthias

Reviewed By: jasongilanfarr, aquamatthias

Subscribers: jenkins, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D9