Validate collection types in RamlGeneratedTypes
Summary:
Previously, if there were any json validation errors in a collection type,
the object(s) would be silently dropped and replaced with the empty collection
Now, they will be validated and only if they are actually missing will an
empty collection take its place.

Test Plan:
Created an app definition with an external volume that was missing required
fields and parsed it into an App - now it is indeed a failure.

Reviewers: jdef, aquamatthias

Reviewed By: jdef, aquamatthias

Subscribers: jenkins, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D200