Set `maxConsecutiveFailures` to max int.
Summary:
Health checks in integration tests would fail because AppMock
would not start fast enough. Since we used to have max failures set to 2
with an interval of one second Marathon would kill the app before it
even started. The integration test don't test the health check so we set
the default to max int. The timeouts of the test will kick in after 90
seconds and kill the app if something is really wrong.

Test Plan: unit-test, integration-test

Reviewers: meichstedt, jasongilanfarr, timcharper

Reviewed By: jasongilanfarr, timcharper

Subscribers: jasongilanfarr, marathon-team, jenkins

Differential Revision: https://phabricator.mesosphere.com/D236