Use one scheduled timer to send TaskStatusUpdates, not one timer for every update.
Summary:
The previous version of this simulator uses the akka scheduler to schedule a message for every task to start.
This version sends every second all messages that are overdue.

Simulate different agents.
Send offers with different agent id.
This will have effect on the offer matching rejection algorithm

Extract errorProbability and set to 0.
The previous version had an error probability of 0.001 which is set to 0 here.

TASK_STAGING
Mesos does not send TASK_STAGING updates, but the previous simulator did.
Removed this to have the same behaviour.

Test Plan: Run main in mesos-simulation and start an app with 50K tasks

Reviewers: jasongilanfarr, meichstedt

Reviewed By: jasongilanfarr, meichstedt

Subscribers: jenkins, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D311