Remove obsolete Unstable MarathonStoreTest
Summary:
The behavior for the state.names() method was to both add a timeout (via
BackToTheFuture.futureToFuture) and return an empty seq for non-fatal
exceptions.

Since the test is testing a special behavior that was removed over a
year ago, and hasn't been modified since, I'm assuming new behavior is
correct and the test is wrong.

Fixes #4840

Test Plan: n/a

Reviewers: jdef, jasongilanfarr, aquamatthias

Reviewed By: jasongilanfarr, aquamatthias

Subscribers: jenkins, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D330