Do not block during reconciliation of health checks.
Summary:
The MarathonHealthCheckManager used the sync version of the InstanceTracker.
More than that it calls this method for every application that has to reconcile.
This PR reconciles all apps in one call, which gives the opportunity to get all current running instances in one call and then start the reconciliation.

After this patch is applied, I did not see any explosion of threads during the scale test.

Test Plan:
Start the simulator and launch 5000 apps with one task.
Watch the number of threads.

Reviewers: meichstedt, janisz, unterstein

Reviewed By: meichstedt, janisz, unterstein

Subscribers: jdef, janisz, jenkins, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D333