Mark task_lost_expunge_gc command line parameter as deprecated.
Summary:
The task_lost_expunge_gc command line parameter does not do anything in 1.4.
Added this information in the command line docs.
Hide the parameter, so --help will not show it any longer, but specification of the parameter will be ignored.

Test Plan:
Read
Try to specify the parameter on the command line
Read marathon --help

Reviewers: jeschkies, unterstein

Reviewed By: jeschkies, unterstein

Subscribers: jenkins, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D343