Create SerialIntegrationTest for running time-sensitive integration tests
Summary:
As documented in #4736, there is sufficient evidence to support a
hyopthesis that a primary reason why the test suite
GracefulTaskKillIntegrationTest (and others) fail, is due to resource
contention caused by other tests running, and the fact that the test
is very time dependent.

More optimally, this test could be re-written to only assert that the
KillPolicy is well formed (and leave it to mesos to verify it's
guarantees that the policy is interpretted as advertised).

Invoking `integration:test` will cause the serial tests to run first,
followed by the parallel tests.

Closes #4736

Test Plan:
It's taken about 4-6 runs in the past to produce a failure for
GracefulTaskKillIntegrationTest. If `integration:test` passes 6 times in a row,
I'd consider that ample enough evidence to support a merge.

Reviewers: jdef, jasongilanfarr

Reviewed By: jdef, jasongilanfarr

Subscribers: marathon-team

Differential Revision: https://phabricator.mesosphere.com/D348