Test if second Marathon fails when it binds to same port.
Summary:
This is tricky. Sometimes all tests would pass but sbt would still report an
error because a test has exited with 137. As it turns out the Zookeepr client
time out would keep the test process alive until it halted.

Moreover the test did not kill all its threads. This fix starts Marathon in a
separate process and destroys it if it times out.

Test Plan: integration-test

Reviewers: timcharper, meichstedt, jasongilanfarr

Subscribers: jenkins, marathon-team