Make sure the HistoryActor only stores task failures once.
Summary:
With introducing the instance abstraction, the HistoryActor started to handle both InstanceChanged events and MesosStatusUpdateEvents for storing TaskFailures. This leads to failures being stored twice if both events are translated to a TaskFailure (the second write overwrites the first because we only keep the latest TaskFailure). Mid-term, handling the MesosStatusUpdateEvent should be removed in favor of InstanceChanged, but the current implementation re provided data is not sufficient to do so and requires more changes. This patch prevents the actor from storing the failure twice by going back to handling MesosStatusUpdateEvent onl< for the time being.

Related to #4792

Test Plan: sbt test

Reviewers: timcharper, aquamatthias, jasongilanfarr

Reviewed By: aquamatthias, jasongilanfarr

Subscribers: jenkins, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D411