Test if second Marathon fails when it binds to same port to fix #4929.
Summary:
This is tricky. Sometimes all tests would pass but sbt would still report an
error because a test has exited with 137. As it turns out the Zookeeper client
timeout would keep the test process alive until it halted.

Moreover the test did not kill all its threads. This fix starts Marathon in a
separate process and destroys it if it times out.

Test Plan: integration-test

Reviewers: meichstedt, jasongilanfarr, zen-dog, timcharper

Reviewed By: jasongilanfarr, timcharper

Subscribers: marathon-team, jenkins

Differential Revision: https://phabricator.mesosphere.com/D404