Create master pipeline based on Jenkinsfile (#4946)
Summary:
The Jenkinsfile defines two setup stages, a compile, test and integration test stage. The final step archives the binaries. Each step updates the commit status. There is a final status when everything passed.

So far we don't have Cobertura support in pipelines so we archive the results. We may just upload the artifacts to S3 and have DataDog create statistics for us.

So far the stages do not mark the commits as pending.

Test Plan:
https://jenkins.mesosphere.com/service/jenkins/view/Marathon/job/marathon-pipelines/job/PR-4946/ 

Reviewers: aquamatthias, janisz, jasongilanfarr, kensipe

Subscribers: jenkins, marathon-team