Fix a race condition (and Replace InstanceKillProgressActor with Akka stream)
Summary:
Karsten and I worked on this patch together. A race condition existed
in KillServiceActor in which an instance could be killed and status
update received before InstanceKillProgressActor subscribed to the
event bus.

The Akka Stream solution is arguably easier to reason about, and,
ensures that the event bus subscription is registered before any
kill operations are sent to Meosos.

Also-By: Karsten Jeschkies <jeskar@web.de>

Test Plan: scale up and down several times

Reviewers: jasongilanfarr, meichstedt, jenkins

Reviewed By: jasongilanfarr, meichstedt

Subscribers: jdef, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D510