remove shutdown hooks; detect shutdown and circumvent reconnect attempts
Summary:
Also, we stop creating actorSystems as default parameters to a method,
and make it such that the entity responsible for creating an actor
system is also responsible for shutting it down.

By detecting shutdown state and circumventing attempts that would try
and reconnect to Zookeeper, we fix infinite-reconnect loops which
prevent clean shutdown and can produce a lot of log messages.

Test Plan: unit tests

Reviewers: jdef, jenkins, jasongilanfarr

Reviewed By: jdef, jenkins, jasongilanfarr

Subscribers: marathon-team

Differential Revision: https://phabricator.mesosphere.com/D507