Await app scale and mark AppDeployIntegrationTest as stable to fix #5019.
Summary:
The `kill all tasks of an App with scale` test in `AppDeployIntegrationTest`
would fail ~2 out of 10 times on my machine. It turns out that
`SchedulerActions.scale` would be called twice. The `withLockFor(runSpecId)`
should have prevented this. The lock is freed when `TasksKilled` is sent to
`self`. This should happen after the future returned by `SchedulerActions.scale`
completes. However, the async/await block returned a `Future[Future[Unit]]`
instead of a `Future[Unit]`. So the lock was freed before the scale completed.

The type checker cannot catch this error because

```
val res: Future[Unit] = Future(Future(Unit))
```

is valid. That's why `Unit` was changed to `Done`. Now the compiler gives an
error.

Test Plan: integration-test

Reviewers: unterstein, meichstedt, jasongilanfarr, timcharper

Reviewed By: jasongilanfarr, timcharper

Subscribers: jenkins, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D462