Improve performance of zookeeper layer and groups
We unnecessarily stringify every protobuf that we serialize to
Zookeeper. This is quite expensive. We modify the code to only serialize
the protobuf if an exception is actually thrown.

Group.toProto unnecessarily creates a few new interim sets, which causes
each proto to hashed unnecessarily. This is wasteful and quite
expensive. Also, since groups are immutable, we cache this result with a
lazy val.