Close curator client in leaderelection.
Summary:
This patch should fix #5036. The shutdown would sometimes end up in an endless
loop for tests because the retry policy is triggering a shutdown but the curator
client is not closed. The error on closing the leader latch triggers the retry
policy again as long as the client is not disconnected.

See the GitHub issue #5036 for details.

Test Plan: integration-test

Reviewers: timcharper, aquamatthias, jasongilanfarr, jenkins

Reviewed By: timcharper, jasongilanfarr, jenkins

Subscribers: jenkins, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D491