ForceExpunge for missing task is a no-op, rather than a failure
Summary:
If a resident app is deleted before Marathon has a chance to
release the associated resources, then later attempts to release them
fail because because we ForceExpunge the task-ID when processing our
UnreserveAndDestroyVolumes operation.

This change makes ForceExpunge for a non-existent task respond with
no-op, rather than failure.

Fixes #5142

Backport of f441c57

Test Plan: n/a

Reviewers: jasongilanfarr, meichstedt

Subscribers: marathon-team

Differential Revision: https://phabricator.mesosphere.com/D547