Find Marathon cap by scaling incremantally.
Summary:
This is the first simplified scale test for finding the maximum number
of instances per app. We start with a simple sleep app and add 1000
instances step by step. The maximum wait time for a scaling to finish is
hard coded to 10 minutes.

A frist run on a 30 nodes cluster with a total of 124 CPUs and 424GB of
memory should that some tasks become unreachable when scaling from
20,000 to 21,000 instances. The 10 minute is reached.

The resources should be enough for more than 53,000 intances.

Test Plan: manual

Reviewers: aquamatthias, kensipe, jenkins, jasongilanfarr

Reviewed By: kensipe, jenkins

Subscribers: marathon-team, jenkins

Differential Revision: https://phabricator.mesosphere.com/D530