Improve Raml Type Generation for discriminator styles.
Summary:
I don't really like discriminator (when it can be parsed unambiguously), but
the early jobs designs had a good case for it, so why not.

(a) Make sure playJsonFormat is (almost) always available.
(b) Jump through the necessary hoops to generate a RAML that Matthias E sent me.

Test Plan: Matthias E's stuff patch now compiles with it.

Reviewers: jdef, timcharper, meichstedt, jenkins

Reviewed By: jdef, meichstedt, jenkins

Subscribers: marathon-team

Differential Revision: https://phabricator.mesosphere.com/D551