Add a custom constructor to ServiceMockFacade and make the default private.
Summary:
The tests previously used the facade as soon as a task was running and did not realize when the task failed (e.g. when failing to bind to the configured port). The custom constructor ensures that the facade is only used when the task actually responds to a request.

Addresses MARATHON-1840, MARATHON-6993

Test Plan: integration:test

Reviewers: timcharper, zen-dog, jenkins, unterstein, jeschkies

Reviewed By: zen-dog, jenkins, unterstein

Subscribers: marathon-team

Differential Revision: https://phabricator.mesosphere.com/D589