Add a custom constructor to ServiceMockFacade and make the default private. (#5319)
Summary:
The tests previously used the facade as soon as a task was running and did not realize when the task failed (e.g. when failing to bind to the configured port). The custom constructor ensures that the facade is only used when the task actually responds to a request.

Note: RestartIntegrationTest is not applicable to releases/1.4 since dependent test classes are not present.

cherry-picked from master 8814cda7439e117adcd7caa34894c321fb0bfe74

Addresses MARATHON-1840, MARATHON-6993, MARATHON-7000

Test Plan: integration:test

Reviewers: timcharper, zen-dog, jenkins, unterstein, jeschkies

Reviewed By: zen-dog, jenkins, unterstein

Subscribers: marathon-team

Differential Revision: https://phabricator.mesosphere.com/D589

Conflicts:
	src/test/scala/mesosphere/marathon/integration/ReadinessCheckIntegrationTest.scala
	src/test/scala/mesosphere/marathon/integration/RestartIntegrationTest.scala