scale graphs
Summary:
code necessary to graph the data from scale testing.

In the language of matplot, there is 1 fig with 2 plots.   The first plot is the time it takes to get to scale.   It includes a desc on the Y axis.   The X axis should list the scales (1, 10, 100, etc) and it does when there is only 1 plot.  When we switch to 2 plots I have been unable to get that to show correctly and it shows up on the 2nd plot.    Under the X axis of the first plot are the details of resources, the number of nodes, cpu and mem.      When a scale attempt fails there is a text at the point of time that failed that labels the max scale reached and what time.   There is also a legend in the top left which helps identify the different tests.   The title of the figure includes the version of marathon under test.

The 2nd plot has a list of errors.   It is a simple dot plot of errors for each test.

Test Plan: run tests

Reviewers: jenkins, jeschkies, zen-dog

Reviewed By: jenkins, zen-dog

Subscribers: jgehrcke, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D609