Homogenized the tests for root and mom
Summary:
This is a refactor of the tests such that more of the tests are run against root and mom mararthons.  This refactor supports the writing of the test once
in the marathon_common_tests.py.   All tests there will be run on root and mom.  The trick is the pytestmark = [pytest.mark.usefixtures('mom_fix')] which puts a
mom fixture on all tests in the test_marathon_on_marathon.py.   All tests which were mom specific or root specific were moved their specific test file.  The names
of the test files were modified to be more clear and consistent.

There are 2 exceptions.   All the pod tests are still in a pod file and included for root testing.  The same is true for dcos service or framework tests.

The remaining tests are:

test_marathon_on_marathon.py - tests mom
test_marathon_on_marathon_ee.py - tests mom ee
test_marathon_root.py - tests root
test_marathon_universe.py - test universe (which is a root test but seems strongly different in some ways)

There are a significant number of tests which were only tested on mom or only tested on root which are now tested on both.

Test Plan: run the tests

Reviewers: zen-dog, jenkins, unterstein

Reviewed By: zen-dog, jenkins

Subscribers: marathon-dev

JIRA Issues: MARATHON_EE-1422

Differential Revision: https://phabricator.mesosphere.com/D679