Adding pinger relay test for local networks
Summary:
This test runs a pinger app and a relay app.  It retrieves the python app from the master via the new
http service (which will be moving into shakedown).   Then a curl call to the relay will invoke a call
to the 2nd pinger app and return back pong to the relay then back to curl.

Follow up tests will test this from a bridge and from the dcos network.

Test Plan: run test

Reviewers: zen-dog, jeschkies, unterstein, jenkins

Reviewed By: zen-dog

Subscribers: marathon-dev

JIRA Issues: MARATHON_EE-1420

Differential Revision: https://phabricator.mesosphere.com/D695