Introduce TTY type and make it available on app and pod container.
Summary:
Raml type TTY added and referenced from App and Pod MesosContainer.
Added a tty to the ServiceDefinition in the zk model.
Added proto reader/writer for TTY.
Handle the tty in TaskBuilder for app and TaskGroupBuilder for pod.
Added test cases for the change.

In order to implement this change an update to Mesos 1.2 was needed (will extract an extra Patch for this).
Sadly Mesos 1.2 breaks framework registration with roles.
See: CORE-977

Test Plan:
sbt test
sbt integration:test

Reviewers: unterstein, timcharper, jenkins

Reviewed By: unterstein, timcharper, jenkins

Subscribers: marathon-dev

JIRA Issues: MARATHON-7062

Differential Revision: https://phabricator.mesosphere.com/D683