Close curator client and suicide on connection lost.
Summary:
This addresses the deadlock we see on some tests involving Marathon shutdowns
and connection loss to ZooKeeper. Curator and the ZooKeeper client try to
reconnect if the connection is lost. This blocks the shutdown on the JVM for
tests at least.

To avoid a race condition between the client closing and a new operation we
syncronize `RichCuratorFramework.build` which should be fine since we return a
Future. This is akin to the implementation in `CuratorElectionService`. At some
point we might want to move to a single curator client instance.

Test Plan: integration-test, loop

Reviewers: aquamatthias, timcharper, jenkins, jasongilanfarr

Reviewed By: aquamatthias, jenkins, jasongilanfarr

Subscribers: marathon-dev, marathon-team, jenkins

JIRA Issues: MARATHON-7184, MARATHON-7230

Differential Revision: https://phabricator.mesosphere.com/D674