Backup & Restore storage version
Summary:
The storage version needs to be handled separately, since it is handled as meta data of the persistence store.
It is necessary to store the version as part of the backup, to ensure later migrations run successfully.

Test Plan:
Backup data from zk or mem store.
Make sure the tar file contains a file called internal.storage.version
Restore and make sure the version is restored successfully.

Reviewers: jasongilanfarr, unterstein, jenkins

Reviewed By: jasongilanfarr, unterstein, jenkins

Subscribers: marathon-dev

JIRA Issues: MARATHON-7266

Differential Revision: https://phabricator.mesosphere.com/D726