Replace Mesos install shell script with Scala script.
Summary:
The current script is some inline Bash code in a Jenkinsfile. The unstable loops
failed because the were not updated. The new script should replace the Bash
code. In the long run we want to replace most of the Jenkinsfile with Scala
scripts.

Test Plan: n/a pipelines did not work.

Reviewers: timcharper, jasongilanfarr, jenkins, unterstein

Reviewed By: jasongilanfarr, jenkins, unterstein

Subscribers: jenkins, marathon-dev, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D706