Enable bridge networking for pods.
Summary:
In the current implementation it is possible to configure a bridged container network with `container/bridge`, but when building the TaskInfo's for pods these networks are ignored and not passed to mesos. This fix changes the behavior to enable bridged networks for pods.

Deferred work:
- https://jira.mesosphere.com/browse/MARATHON-7398
- https://jira.mesosphere.com/browse/MARATHON-7414

Test Plan: sbt test

Reviewers: zen-dog, timcharper, jdef, jenkins, jeschkies, kensipe

Reviewed By: zen-dog, jdef, jenkins, jeschkies, kensipe

Subscribers: jeschkies, kensipe, marathon-dev, marathon-team

JIRA Issues: MARATHON-7204

Differential Revision: https://phabricator.mesosphere.com/D801