Do not block on purge in MarathonSchedulerActor. (#5411)
Summary:
The synchronous call to purge can block up to 33 seconds. This caused a
flakiness in GroupDeployIntegration tests. The MarathonSchedulerActor
was not able to process other message and thus the Deploy command was
not received.

This does not solve the issue of the delay withing the purge. See
MARATHON-7365 for further investigation.

Test Plan: pipeline

Reviewers: timcharper, zen-dog, jenkins

Reviewed By: timcharper, zen-dog, jenkins

Subscribers: marathon-dev, marathon-team, jenkins

JIRA Issues: MARATHON-7430

Differential Revision: https://phabricator.mesosphere.com/D814