Changed the position of the legend in the graph.
Summary: Changed the position of the legend in the graph. Now the legends are outside their respective graphs and not inside which makes the results look neater and the graph does not look cluttered.

Test Plan: Tested with the example data to see how the changes in the graph looks like.

Reviewers: kensipe, jenkins

Reviewed By: kensipe, jenkins

Subscribers: marathon-dev

Differential Revision: https://phabricator.mesosphere.com/D870