Jepsen project to automate testing of various Marathon operations in a cluster in the presence of network splits
Summary:
This is a Jepsen project that will test variors Marathon operations. Currently, I have created a basic skeleton for setting up
Marathon, Mesos and Zookeeper.

Test Plan: Test out if Marathon, Mesos and Zookeeper as configured with adding a noop for teardown once the skeleton is filled up. Currently testing with the printing infos.

Reviewers: ichernetsky, jenkins

Reviewed By: ichernetsky, jenkins

Subscribers: marathon-dev

Differential Revision: https://phabricator.mesosphere.com/D854