Remove WorkQueue lock use
Summary:
The ThreadDump reported for MARATHON-7400 reveals that a deadlock
occurs when trying to access the mutable WorkQueue HashMap.

We are unable to explain why this happens.

There is no need to rely on custom concurrency logic here. TrieMap
provides sufficient concurrency features.

Note that in this change we assume that WorkQueue.apply is threadsafe

Test Plan: n/a

Reviewers: jeschkies, zen-dog, jenkins

Reviewed By: jeschkies, zen-dog, jenkins

Subscribers: jdef, marathon-dev

JIRA Issues: MARATHON-7400

Differential Revision: https://phabricator.mesosphere.com/D863