Functions for launching and terminating Marathon Framework via Jepsen added
Summary: Added functions which brings up marathon framework, talks to Mesos master, and terminates it during the teardown phase. Also, changed the clojure format tool based on Ivan's suggestion as this tool is more user-friendly.

Test Plan: Check if the plumbing is correct by shutting the teardown phase off. In future, a sample task will be created via Jepsen which if launches via Marathon correctly on the VMs, will verify the plumbing to be correct.

Reviewers: ichernetsky, kensipe, jenkins

Reviewed By: ichernetsky

Subscribers: marathon-dev

Differential Revision: https://phabricator.mesosphere.com/D884