Small debian packages update fix which now makes sure that installation of zookeeper does not fail
Summary: Zookeeper installation on the VMs via Jepsen started failing leading to the Jepsen setup and teardown lifecycle to terminate. Reason was that there was a 404 error while installing zookeeper which was not happening before. Hence I added a debian update jepsen command before zookeeper installation which fixed the issue.

Test Plan: Test if the setup and teardown lifecycle completes successfully.

Reviewers: ichernetsky, timcharper, jenkins

Reviewed By: ichernetsky, timcharper

Subscribers: marathon-dev

Differential Revision: https://phabricator.mesosphere.com/D887