Remove poorly performing implicit java wrappers
Summary:
These implicit methods led to poor performnance surprising
behavior. For example, the Java Map proxy implementation led to the
creation of a brand new immutable Scala set and re-hashing of all
underlying data in order to perform some operation, only to have said
data structure immediately discarded.

The Set implementation was not much better; it gave the appearance of
being a set but caused all set operations to be done by sequential
scans.

It is far better to be explicit about these conversions when they
could have potential performance pitfalls such as this.

Test Plan: n/a

Reviewers: jeschkies, zen-dog, jenkins

Reviewed By: jeschkies, zen-dog, jenkins

Subscribers: jdef, marathon-dev

Differential Revision: https://phabricator.mesosphere.com/D864